requirements.txt
# Used Python 3.10
chardet==3.0.4
nltk==3.8.1
numpy==1.26.3
pandas==1.5.3
xmltodict==0.12.0
setuptools==65.5.0
PyMuPDF==1.23.15
python-docx==1.1.0